move-formatAllTimeDate-function-to-utils.ts-and-import-in-CoinProfit.tsx
