+++ src/app/utils.ts
@@ -16,3 +16,19 @@ export const formatCurrencyWithCommas = (num: number, currency: string) => {
 export const formatNumberWithCommas = (num: number) => {
   return new Intl.NumberFormat().format(num);
 };
+
+export const formatAllTimeDate = (dateStr: string): string => {
+  const date = new Date(dateStr);
+  const options: Intl.DateTimeFormatOptions = {
+    weekday: "short",
+    day: "numeric",
+    month: "short",
+    year: "numeric",
+    hour: "numeric",
+    minute: "numeric",
+    second: "numeric",
+    timeZoneName: "short",
+  };
+  const formatter = new Intl.DateTimeFormat("en-US", options);
+  return formatter.format(date);
+};
